refactor: add controller layer
